Twitter_Bot_Final/requirements.txt
boto3==1.15.0
dataframe_image==0.1.1
numpy==1.18.1
pandas==1.2.0
Pillow==9.2.0
requests==2.28.1
tweepy==4.10.0
matplotlib==3.4
